+++ packages/orq-rc/src/orq_ai_sdk/evals.py
@@ -3,10 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    createevalop as models_createevalop,
-    updateevalop as models_updateevalop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -214,10 +210,7 @@ class Evals(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createevalop.CreateEvalRequestBody,
-                models_createevalop.CreateEvalRequestBodyTypedDict,
-            ]
+            Union[models.CreateEvalRequestBody, models.CreateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -314,10 +307,7 @@ class Evals(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createevalop.CreateEvalRequestBody,
-                models_createevalop.CreateEvalRequestBodyTypedDict,
-            ]
+            Union[models.CreateEvalRequestBody, models.CreateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -415,10 +405,7 @@ class Evals(BaseSDK):
         *,
         id: str,
         request_body: Optional[
-            Union[
-                models_updateevalop.UpdateEvalRequestBody,
-                models_updateevalop.UpdateEvalRequestBodyTypedDict,
-            ]
+            Union[models.UpdateEvalRequestBody, models.UpdateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -524,10 +511,7 @@ class Evals(BaseSDK):
         *,
         id: str,
         request_body: Optional[
-            Union[
-                models_updateevalop.UpdateEvalRequestBody,
-                models_updateevalop.UpdateEvalRequestBodyTypedDict,
-            ]
+            Union[models.UpdateEvalRequestBody, models.UpdateEvalRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,